requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
aiomysql==0.2.0
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymysql==1.1.0
